+++ core/src/avm1/globals.rs
@@ -4,6 +4,7 @@ use crate::avm1::function::{Executable, FunctionObject};
 use crate::avm1::property::Attribute;
 use crate::avm1::property_decl::{define_properties_on, Declaration};
 use crate::avm1::{Object, ScriptObject, TObject, Value};
+use crate::display_object::{DisplayObject, Lists, TDisplayObject, TDisplayObjectContainer};
 use crate::string::{AvmString, WStr, WString};
 use gc_arena::Collect;
 use gc_arena::MutationContext;
@@ -26,7 +27,6 @@ pub(crate) mod context_menu_item;
 pub mod convolution_filter;
 mod date;
 pub mod displacement_map_filter;
-pub(crate) mod display_object;
 pub mod drop_shadow_filter;
 pub(crate) mod error;
 mod external_interface;
@@ -1243,6 +1243,51 @@ pub fn create_globals<'gc>(
         globals.into(),
         broadcaster_functions,
     )
+}
+
+/// Depths used/returned by ActionScript are offset by this amount from depths used inside the SWF/by the VM.
+/// The depth of objects placed on the timeline in the Flash IDE start from 0 in the SWF,
+/// but are negative when queried from MovieClip.getDepth().
+/// Add this to convert from AS -> SWF depth.
+const AVM_DEPTH_BIAS: i32 = 16384;
+
+/// The maximum depth that the AVM will allow you to swap or attach clips to.
+/// What is the derivation of this number...?
+const AVM_MAX_DEPTH: i32 = 2_130_706_428;
+
+/// The maximum depth that the AVM will allow you to remove clips from.
+/// What is the derivation of this number...?
+const AVM_MAX_REMOVE_DEPTH: i32 = 2_130_706_416;
+
+fn get_depth<'gc>(
+    activation: &mut Activation<'_, 'gc, '_>,
+    this: Object<'gc>,
+    _args: &[Value<'gc>],
+) -> Result<Value<'gc>, Error<'gc>> {
+    if let Some(display_object) = this.as_display_object() {
+        if activation.swf_version() >= 6 {
+            let depth = display_object.depth().wrapping_sub(AVM_DEPTH_BIAS);
+            return Ok(depth.into());
+        }
+    }
+    Ok(Value::Undefined)
+}
+
+pub fn remove_display_object<'gc>(
+    this: DisplayObject<'gc>,
+    activation: &mut Activation<'_, 'gc, '_>,
+) {
+    let depth = this.depth().wrapping_sub(0);
+    // Can only remove positive depths (when offset by the AVM depth bias).
+    // Generally this prevents you from removing non-dynamically created clips,
+    // although you can get around it with swapDepths.
+    // TODO: Figure out the derivation of this range.
+    if depth >= AVM_DEPTH_BIAS && depth < AVM_MAX_REMOVE_DEPTH && !this.removed() {
+        // Need a parent to remove from.
+        if let Some(mut parent) = this.avm1_parent().and_then(|o| o.as_movie_clip()) {
+            parent.remove_child(&mut activation.context, this, Lists::all());
+        }
+    }
 }
 
 #[cfg(test)]
