+++ core/src/avm1/globals/video.rs
@@ -2,7 +2,6 @@
 
 use crate::avm1::activation::Activation;
 use crate::avm1::error::Error;
-use crate::avm1::globals::display_object;
 use crate::avm1::object::Object;
 use crate::avm1::value::Value;
 use crate::avm1::ScriptObject;
@@ -20,11 +19,8 @@ pub fn constructor<'gc>(
 pub fn create_proto<'gc>(
     gc_context: MutationContext<'gc, '_>,
     proto: Object<'gc>,
-    fn_proto: Object<'gc>,
+    _fn_proto: Object<'gc>,
 ) -> Object<'gc> {
     let object = ScriptObject::object(gc_context, Some(proto));
-
-    display_object::define_display_object_proto(gc_context, object, fn_proto);
-
     object.into()
 }
